requirements.txt
tweepy==3.7.0
requests==2.22.0
python-dotenv==0.10.3